I moved the tracker into the page's userdata
I was storing the tracker in the extension's memory, which
caused problems because the caceobjs stored in the tracker
had references to elements on the pages.  When the page
was closed, the memory wouldn't be deallocated because of
the tracker.

I also disabled the cleanCacheObj since it is now auto-cleaned
when the document is unloaded.

See also: https://bugzilla.mozilla.org/show_bug.cgi?id=672619